Angle: Add Angle as a backend for rendering
This commit sets Angle to be the default backend used during rendering, however it doesn't change the compositing of the UnoHost. An experiment testing out shared texture between the tool and the UnoHost can be toggled in UnoHostControlImplementation.cs, by setting `UseWindowBasedSurface = false`. Generally using Angle as a backend feels smoother, and there is less rendering glitches in my experience.