Windows: make all processes run in 64-bit mode
Before we had some processes running in 32-bit mode while the main process
is running in 64-bit mode.

In Fuselibs 1.10.0-rc1 we longer have a 32-bit version of
Fuse.Text-PInvoke.dll, and running in 32-bit mode results in corrupt
viewports when an app crashes on start up.

This commit fixes that by making all processes run in 64-bit mode.