Stuff: replace uno.exe
The old uno.exe had some strange behaviour where it ends up downloading an
old version of Uno 1.8.0 from NuGet when building the preview app.

This is not expected behaviour, and we want to build the preview app using
the same version of Uno that we ship with Fuse Studio.

If we take uno.exe from Fuselibs and change .unopath to point to the
packages/ directory, we link to the same Uno that were downloaded while
building the Fuse Studio solution.